Huuan

+++ client/src/components/pages/InternshipPage.js
@@ -3,6 +3,8 @@ import React from 'react'
 import {
   Row, Col, Card, CardBody, MDBIcon, MDBModalBody, MDBInput, MDBBtn, MDBModal,
 } from 'mdbreact';
+import ReactNotification from "react-notifications-component";
+import "react-notifications-component/dist/theme.css";
 import TextField from '@material-ui/core/TextField';
 import MUIDataTable from "mui-datatables";
 import DatePickers from './sections/DatePickers'
@@ -19,7 +21,7 @@ import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import Button from '@material-ui/core/Button';
 import './attendance.css';
-
+import $ from 'jquery';
 /* Import MUIDataTable using command "npm install mui-datatables --save" */
 
 function TabContainer(props) {
@@ -50,6 +52,8 @@ class InternPage extends React.Component {
 
   constructor() {
     super();
+    this.addNotification = this.addNotification.bind(this);
+    this.notificationDOMRef = React.createRef();
     this.state = {
       modal: false,
       data: [],
@@ -58,8 +62,66 @@ class InternPage extends React.Component {
       courseList: [],
       isUpdate: false,
       checkValidate: true,
-      courseName: ''
+      CourseName: ''
     };
+  }
+
+  addNotification(kind) {
+    switch (kind) {
+      case "successAdd":
+        this.notificationDOMRef.current.addNotification({
+          title: "Success",
+          message: "Add course successfully !",
+          type: "success", //success, danger, default, info, warning or custom
+          message: "This is a success message!",
+          insert: "top",
+          container: "top-right",
+          animationIn: ["animated", "fadeIn"],
+          animationOut: ["animated", "fadeOut"],
+          dismiss: { duration: 2000 },
+          dismissable: { click: true }
+        });
+        break;
+      case "errorAdd":
+        this.notificationDOMRef.current.addNotification({
+          title: "Error",
+          message: "Add course fail",
+          type: "danger",
+          insert: "top",
+          container: "top-right",
+          animationIn: ["animated", "fadeIn"],
+          animationOut: ["animated", "fadeOut"],
+          dismiss: { duration: 2000 },
+          dismissable: { click: true }
+        });
+        break;
+      case "successUpdate":
+        this.notificationDOMRef.current.addNotification({
+          title: "Success",
+          message: "Update course successfully !",
+          type: "success", //success, danger, default, info, warning or custom
+          insert: "top",
+          container: "top-right",
+          animationIn: ["animated", "fadeIn"],
+          animationOut: ["animated", "fadeOut"],
+          dismiss: { duration: 2000 },
+          dismissable: { click: true }
+        });
+        break;
+      case "successDelete":
+        this.notificationDOMRef.current.addNotification({
+          title: "Success",
+          message: "Delete course successfully !",
+          type: "success", //success, danger, default, info, warning or custom
+          insert: "top",
+          container: "top-right",
+          animationIn: ["animated", "fadeIn"],
+          animationOut: ["animated", "fadeOut"],
+          dismiss: { duration: 2000 },
+          dismissable: { click: true }
+        });
+        break;
+    }
   }
 
   GetInternList() {
@@ -70,24 +132,22 @@ class InternPage extends React.Component {
         let NewData = []
         let stt = 1
         data.map(row => {
-          NewData.push([stt, row.Intern.ID, row.Intern.Name, row.Intern.PhoneNumber, row.Intern.Email, row.Intern.Gender ? "Male" : "Female", 
-          (new Date(row.Intern.DoB)).toLocaleDateString('en-US', options),
-           row.Intern.University, row.Intern.Faculty, row.Course,
+          NewData.push([stt, row.Intern.ID, row.Intern.Name, row.Intern.PhoneNumber, row.Intern.Email, row.Intern.Gender ? "Male" : "Female",
+            (new Date(row.Intern.DoB)).toLocaleDateString('en-US', options),
+            row.Intern.University, row.Intern.Faculty, row.Course,
             // format datetime,
-            row.Department, row.CourseID])
+            ])
           stt++
-          return NewData
+          return NewData,
+          console.log(NewData)
         })
         this.setState({
           internList: NewData
         })
-        console.log(data)
+        // console.log(data)
       });
   }
 
-  handleChangeTab = (event, value) => {
-    this.setState({ value });
-  };
 
   componentDidMount() {
 
@@ -98,11 +158,9 @@ class InternPage extends React.Component {
 
   toggleIntern = () => {
     this.setState({
-      modalIntern: !this.state.modalIntern,
+      modalIntern: !this.state.modalIntern,/*false*/
     });
   };
-
-
 
   addIntern = () => {
     this.setState({
@@ -110,7 +168,7 @@ class InternPage extends React.Component {
       phone: "",
       email: "",
       gender: "",
-      courseName: "",
+      CourseName: "",
       dob: "",
       University: "",
       Faculty: "",
@@ -122,11 +180,13 @@ class InternPage extends React.Component {
   }
 
   handlerAddIntern = () => {
+    // if (confirm("You definitely want to add intern?")) { //eslint-disable-line
+
+    // }
     var moment = require('moment');
-    
+
     const date = moment.utc(this.state.dob).format();
-    // const dt = this.state.dob.split(/-|\s/)
-    // let date = new Date(dt[2], dt[1], dt[0])
+
     const data = {
       "Name": this.state.name,
       "PhoneNumber": this.state.phone,
@@ -139,18 +199,29 @@ class InternPage extends React.Component {
       "IsDeleted": false
     }
     console.log(data)
-    fetch("http://localhost:8080/intern",
-      {
-        method: "POST",
-        mode: "no-cors",
-        headers: {
-          "Content-Type": "application/json"
-        },
-        body: JSON.stringify(data)
-      })
-      .then(this.GetInternList())
+    $.ajax({
+      url: "http://localhost:8080/intern",
+      type: "POST",
+      dataType: "json",
+      data: JSON.stringify(data),
+      success: function( response ) {
+        
+      }
+    });
+    
     this.toggleIntern()
-
+    // fetch("http://localhost:8080/intern",
+    //   {
+    //     method: "POST",
+    //     mode: "no-cors",
+    //     headers: {
+    //       "Content-Type": "application/json"
+    //     },
+    //     body: JSON.stringify(data)d
+    //   })
+    //   .then(this.GetInternList())
+    // this.toggleIntern()
+    // this.addNotification("successAdd")
     // window.location.reload();
   }
 
@@ -171,27 +242,32 @@ class InternPage extends React.Component {
   }
 
   handlerDeleteIntern = () => {
-    fetch("http://localhost:8080/intern/" + this.state.id, {
-      method: 'DELETE',
-      mode: 'cors'
-    })
-      .then(this.GetInternList())
-    this.toggleIntern()
-
+    if (confirm("bạn chắc chắn muốn xóa ?")) { //eslint-disable-line
+      fetch("http://localhost:8080/intern/" + this.state.id, {
+        method: 'DELETE',
+        mode: 'cors',
+      })
+        .then(this.GetInternList())
+        
+      this.toggleIntern()
+      alert("OK")
     window.location.reload();
+    }
+    else
+    {
+      this.toggleIntern()
+      alert("FAIL")
+    }
+
+    
   }
 
   handlerEditIntern = () => {
     var moment = require('moment');
-    
+
     const date = moment.utc(this.state.dob).format();
 
-    // const dt = this.state.dob.split(/-|\s/)
-    // let date = new Date(dt[2], dt[1], dt[0])
-
-    //new Date(dt[2], dt[1], dt[0])
-    // console.log(this.state.date)
-    console.log("dt"+date);
+    console.log("dt" + date);
     const data = {
       "Name": this.state.name,
       "PhoneNumber": this.state.phone,
@@ -203,7 +279,8 @@ class InternPage extends React.Component {
       "CourseID": this.state.course,
       "IsDeleted": false
     }
-    fetch("http://localhost:8080/internu/" + this.state.id,{
+    console.log(data)
+    fetch("http://localhost:8080/internu/" + this.state.id, {
       method: 'PUT',
       mode: 'cors',
       headers: {
@@ -215,44 +292,10 @@ class InternPage extends React.Component {
     this.toggleIntern()
     console.log(this.state.id)
 
-    window.location.reload();
+    this.addNotification("successUpdate")
+
+    // window.location.reload();
   }
-
-  // handlerEditIntern = (e) => {
-  //   e.preventDefault();
-
-  //   var moment = require('moment');
-
-  //   const date = moment.utc(this.state.dob).format();
-
-  //   // const dt = this.state.dob.split(/-|\s/)
-  //   // let date = new Date(dt[2], dt[1], dt[0])
-  //   const data = {
-  //     "Name": this.state.name,
-  //     "PhoneNumber": this.state.phone,
-  //     "Email": this.state.email,
-  //     "Gender": this.state.gender === "Male" ? true : false,
-  //     "DoB": date,
-  //     "University": this.state.University,
-  //     "Faculty": this.state.Faculty,
-  //     "CourseID": this.state.course,
-  //     "IsDeleted": false
-  //   }
-  //   console.log(data)
-  //   fetch("http://localhost:8080/internu/" + this.state.id, {
-  //     method: 'PUT',
-  //     mode: 'cors',
-  //     headers: {
-  //       "Content-Type": "application/json"
-  //     },
-  //     body: JSON.stringify(data)
-  //   })
-  //     .then(this.GetInternList())
-  //   this.toggleIntern()
-  //   console.log(this.state.id)
-
-  //   // window.location.reload();
-  // }
 
 
   columnsIntern = [
@@ -267,7 +310,7 @@ class InternPage extends React.Component {
       name: "ID",
       options: {
         filter: false,
-        sort: false,
+        sort: true,
         display: "excluded"
       }
     },
@@ -371,6 +414,8 @@ class InternPage extends React.Component {
     },
     onRowClick: (rowData, rowState) => {
       let std = this.convertDate(rowData[6])
+      // let courseid = (rowData[9])
+    
       this.setState({
 
         id: rowData[1],
@@ -387,36 +432,45 @@ class InternPage extends React.Component {
         checkValidate: true
       });
       this.toggleIntern()
+      // console.log(this.state.course)
     }
   }
-
+  // convertCourse(courseName){
+  //   fetch("http://localhost:8080/coursename/" + courseName, {
+  //     method: 'GET',
+  //     mode: 'cors',
+  //     headers: {
+  //       "Content-Type": "application/json"
+  //     }
+  //   })
+  // }
   convertDate(rowData) {
     var moment = require('moment')
     let strDate = ""
     let strMon = ""
     let strYea = ""
-      let ye = moment(rowData).get('year');
-      let mo = moment(rowData).get('month') + 1;  // 0 to 11
-      let da = moment(rowData).get('date');
-      if(da < 10)
-        strDate = "0"+da
-      else
-        strDate = ''+da
-      if(mo < 10)
-        strMon = "0"+mo 
-        else 
-        strMon = ''+mo
-      if(ye <1000){
-        strYea = "0" +ye
-        if(ye <100){
-          strYea = "0"+strYea
-          if(ye <10) 
-            strYea = "0"+strYea
-        }
+    let ye = moment(rowData).get('year');
+    let mo = moment(rowData).get('month') + 1;  // 0 to 11
+    let da = moment(rowData).get('date');
+    if (da < 10)
+      strDate = "0" + da
+    else
+      strDate = '' + da
+    if (mo < 10)
+      strMon = "0" + mo
+    else
+      strMon = '' + mo
+    if (ye < 1000) {
+      strYea = "0" + ye
+      if (ye < 100) {
+        strYea = "0" + strYea
+        if (ye < 10)
+          strYea = "0" + strYea
       }
-      else
-        strYea = ''+ye
-      return  strYea+"-"+strMon+"-"+strDate 
+    }
+    else
+      strYea = '' + ye
+    return strYea + "-" + strMon + "-" + strDate
   }
   checkValidate() {
 
@@ -522,6 +576,9 @@ class InternPage extends React.Component {
           <Col md="12">
             <Card className="mt-5">
               <CardBody>
+                <div className="app-content">
+                  <ReactNotification ref={this.notificationDOMRef} />
+                </div>
                 <MDBBtn
                   className="mb-3 blue darken-2"
                   onClick={this.addIntern}>
@@ -547,69 +604,92 @@ class InternPage extends React.Component {
             size="md"
             cascading>
 
-            <MDBModalBody>
+            <MDBModalBody >
               <MDBInput fullwidth="true" size="" label="Name" name="name" value={this.state.name} onChange={this.handleChangeValue.bind(this)} />
               <MDBInput fullwidth="true" label="Phone" name="phone" value={this.state.phone} onChange={this.handleChangeValue.bind(this)} />
               <MDBInput fullwidth="true" label="Email" iconClass="dark-grey" name="email" value={this.state.email} onChange={this.handleChangeValue.bind(this)} />
-              <MDBInput label="Gender" name="gender" value={this.state.gender} onChange={this.handleChangeValue.bind(this)} />
-              <label>Course</label>
+              <FormControl fullWidth>
+                <InputLabel htmlFor="select-multiple">Gender</InputLabel>
+                <Select fullWidth label="Gender" name="gender" value={this.state.gender} onChange={this.handleChangeValue.bind(this)}>
+                  <MenuItem value="Male">Male</MenuItem>
+                  <MenuItem value="Female">Female</MenuItem>
+                </Select>
+
+              </FormControl ><br />
+              <FormControl fullWidth>
+
+                <InputLabel htmlFor="select-multiple">Course</InputLabel>
+                <Select fullWidth label="Course" name="course" value={this.state.course} onChange={this.handleChangeValue.bind(this)}>
+                  {this.state.courseList.map(function (course, index) {
+                    return <MenuItem key={index} value={course.ID}>{course.Name}</MenuItem>;
+                  })}
+                </Select>
+              </FormControl>
+
+              {/* <label>Course</label>
               <select className="browser-default custom-select custom-dropdown custom-margin" onChange={this.handleChangeValue.bind(this)} name="course" size="lg">
                 {this.state.courseList.map(function (course, index) {
                   return <option key={index} value={course.ID}>{course.Name}</option>;
                 })}
-              </select><br/>
+              </select><br /> */}
 
-              {/* <DatePickers
-                label="Day of Birth"
-                name="dob"
+              <MDBInput
+                label="Dob" name="dob" id="date" type="date"
                 value={this.state.dob}
                 onChange={this.handleChangeValue.bind(this)}
-              /> */}
-              <label>Date</label>
-              <input type="date" name="dob" value={this.state.dob}  onChange={this.handleChangeValue.bind(this)}/><br></br>
-              <TextField label="University" name="text" name="University" value={this.state.University} onChange={this.handleChangeValue.bind(this)} />
+                InputLabelProps={{
+                  shrink: true,
+                }}
+              />
+
+              {/* <label>Date</label>
+              <input type="date" name="dob" size="lg" value={this.state.dob} onChange={this.handleChangeValue.bind(this)} /><br></br> */}
+
+              <TextField fullWidth label="University" name="text" name="University" value={this.state.University} onChange={this.handleChangeValue.bind(this)} />
               <div className="text-center mt-1-half">
-                <TextField fullWidth label="Faculty" name="text" name="Faculty" value={this.state.Faculty} onChange={this.handleChangeValue.bind(this)} />
-                <div className="text-center mt-1-half">
-                  {
-                    this.state.isUpdate === false &&
-                    <MDBBtn
-                      className="mb-2 blue darken-2"
-                      onClick={this.handlerAddIntern}
-                    >
-                      Create
+                {/* <TextField fullWidth="true" lable="University" name="text" name="University" value={this.state.University} onChange={this.handleChangeValue.bind(this)} />
+                <div className="text-center mt-1-half"> */}
+                  <TextField fullWidth label="Faculty" name="text" name="Faculty" value={this.state.Faculty} onChange={this.handleChangeValue.bind(this)} />
+                  <div className="text-center mt-1-half">
+                    {
+                      this.state.isUpdate === false &&
+                      <MDBBtn
+                        className="mb-2 blue darken-2"
+                        onClick={this.handlerAddIntern}
+                      >
+                        Create
                   <MDBIcon icon="send" className="ml-1" />
-                    </MDBBtn>
-                  }
-                  {
-                    this.state.isUpdate &&
+                      </MDBBtn>
+                    }
+                    {
+                      this.state.isUpdate &&
 
 
-                    <MDBBtn
-                      className="mb-2 blue darken-2"
-                      onClick={this.handlerEditIntern}>
-                      Update
+                      <MDBBtn
+                        className="mb-2 blue darken-2"
+                        onClick={this.handlerEditIntern}>
+                        Update
                   <MDBIcon icon="edit" className="ml-1" />
-                    </MDBBtn>
-                  }
-                  {
-                    this.state.isUpdate &&
+                      </MDBBtn>
+                    }
+                    {
+                      this.state.isUpdate &&
 
 
-                    <MDBBtn
-                      className="mb-2 blue darken-2"
-                      onClick={this.handlerDeleteIntern}>
-                      Delete
+                      <MDBBtn
+                        className="mb-2 blue darken-2"
+                        onClick={this.handlerDeleteIntern}>
+                        Delete
                   <MDBIcon icon="trash" className="ml-1" />
-                    </MDBBtn>
-                  }
+                      </MDBBtn>
+                    }
+                  </div>
                 </div>
-              </div>
             </MDBModalBody>
           </MDBModal>
         </Row>
       </React.Fragment>
-    )
-  }
-}
+        )
+      }
+    }
 export default withStyles(styles)(InternPage);
